// src/assets/types/constants.tsx
export enum constants {
    CREATE_START = "CREATE_START",
    CREATE_SUCCESS = "CREATE_SUCCESS",
    CREATE_ERROR = "CREATE_ERROR",
    RETRIEVE_START = "RETRIEVE_START",
    RETRIEVE_SUCCESS = "RETRIEVE_SUCCESS",
    RETRIEVE_ERROR = "RETRIEVE_ERROR",
    RETRIEVE_ALL_START = "RETRIEVE_ALL_START",
    RETRIEVE_ALL_SUCCESS = "RETRIEVE_ALL_SUCCESS",
    RETRIEVE_ALL_ERROR = "RETRIEVE_ALL_ERROR",
    UPDATE_START = "UPDATE_START",
    UPDATE_SUCCESS = "UPDATE_SUCCESS",
    UPDATE_ERROR = "UPDATE_ERROR",
    DELETE_START = "DELETE_START",
    DELETE_SUCCESS = "DELETE_SUCCESS",
    DELETE_ERROR = "DELETE_ERROR",
    DELETE_ALL_START = "DELETE_ALL_START",
    DELETE_ALL_SUCCESS = "DELETE_ALL_SUCCESS",
    DELETE_ALL_ERROR = "DELETE_ALL_ERROR",

    CHANGE_PASSWORD_START = "DELETE_START",
    CHANGE_PASSWORD_SUCCESS = "DELETE_SUCCESS",
    CHANGE_PASSWORD_ERROR = "DELETE_ERROR",
}